some modification in signinReducer on login and register but still rendering of AppBarMenue not okay

+++ src/reducers/signinReducer.js
@@ -4,69 +4,65 @@ import * as apiauth from '../api/auth'
 export const signinReducer = (state,action) => {
     switch(action.type){
         case 'REGISTER':
-            registerUser(action.user,action.setMessage,action.history)
-            return 'registered'
+            return registerUser(action.user,action.setMessage,action.history)
         case 'LOGIN':
-            loginUser(action.user,action.setMessage,action.history)
-            return 'signedin'
+            return loginUser(action.user,action.setMessage,action.history);
         case 'LOGOUT':
             logoutUser(action.setMessage,action.history)
             return null
         default:
-            return state
+            return action.history.push('/')
     }
 }
 
-const registerUser = (userData,setMessage,history) => {
-    apiauth.register(userData)
-    .then(response => {
+const registerUser = async (userData,setMessage,history) => {
+    try{
+        const response = await apiauth.register(userData)
         if(response.status === 200){
             console.log(response.status, response.data)
             localStorage.setItem('profile',JSON.stringify(response.data))
             setMessage(`Hello ${response.data.message.username}! Welcome as a new user !`)
-            history.push('/')
-        }
-    })
-    .catch(error => {
+            history.push('/') // return JSON.parse(localStorage.getItem('profile')).message
+        } 
+    } catch(error){
         if(error.response){
-            console.log(error.response.data)
-            if(error.response.data.name === "TokenExpiredError"){
-                localStorage.clear()
-                setMessage('Your session expired, please login again !')
-                history.push('/')
-            } else {
-                setMessage('Ups, something went wrong !')
-                history.push('/')
-            }
+            console.log(error.response)
+            setMessage(error.response.data.message)
+        } else {
+            console.log('error handling')
+            setMessage('Ups, something went wrong !')
         }
-    })
+        return null
+    }
 }
 
 const loginUser = async (userData,setMessage,history) => {
     try{
         const response = await apiauth.login(userData)
         if(response.status === 200){
-            console.log(response.data)
-            localStorage.setItem('profile',JSON.stringify(response.data))
-            // setLoggedInUser(localStorage.getItem('profile') && JSON.parse(localStorage.getItem('profile')).message)
-            setMessage(`Hello ${response.data.message.username}! Welcome back again !`)
-            return response
-            history.push('/')
+            if(response.data.message === 'No User Exists'){
+                console.log('user does not exist')
+                setMessage(`${response.data.message}! Please try again or just watch our nice pictures !`)
+            } else {
+                console.log(response.data)
+                localStorage.setItem('profile',JSON.stringify(response.data))
+                setMessage(`Hello ${response.data.message.username}! Welcome back again !`)
+                history.push('/')//return JSON.parse(localStorage.getItem('profile')).message
+            }
         }
     } catch(error){
+        console.log('HIT LOGIN ERROR')
         if(error.response){
             console.log(error.response.data)
             if(error.response.data.name === "TokenExpiredError"){
                 localStorage.clear()
                 setMessage('Your session expired, please login again !')
-                history.push('/')
-            } else {
-                setMessage('Ups, something went wrong !')
-                history.push('/')
             }
+        } else {
+            setMessage('Ups, something went wrong !')
         }
+        history.push('/')
     }
-
 }
 
 const logoutUser = (setMessage,history) => {
